+++ src/frontend/apps/e2e/__tests__/app-desk/footer.spec.ts
@@ -87,4 +87,13 @@ test.describe('Footer', () => {
       await expect(page).toHaveURL(url);
     });
   }
+
+  test('check if the app version is visible', async ({ page }) => {
+    const footer = page.locator('footer').first();
+    await expect(
+      footer.getByText(
+        'Version: NA • Unless otherwise stated, all content on this site is under',
+      ),
+    ).toBeVisible();
+  });
 });